Updated Functionality
Changed Keyboard shortcut for Settings to mod+, instead of Control, so
on different platforms it will work as intended.

Updated TODO for future implementation of how Projects will work

Fixed about-panel.css So there is 20px of padding on the bottom of the
window

Removed some Code from Drag events.

Added Borders to .new-content in attempt to fix container overlapping
window border.